+++ profiles_api/admin.py
@@ -8,7 +8,7 @@ from profiles_api.question.question_model import Question
 from profiles_api.answer.answer_model import Answer
 from profiles_api.test.test_model import Test
 from profiles_api.completed_test.completed_test_model import CompletedTest
-
+from profiles_api.theory_page.theory_page_model import TheoryPage
 
 admin.site.register(models.UserProfile)
 admin.site.register(models.ProfileFeedItem)
@@ -18,5 +18,5 @@ admin.site.register(Subtopic)
 admin.site.register(Answer)
 admin.site.register(Test)
 admin.site.register(CompletedTest)
-admin.site.register(models.TheoryPage)
+admin.site.register(TheoryPage)
