+++ profiles_api/theory_page/theory_page_api_view.py
@@ -1,0 +1,24 @@
+from rest_framework.views import APIView
+from rest_framework.response import Response
+from rest_framework import viewsets
+from rest_framework.authentication import TokenAuthentication
+from rest_framework.permissions import IsAuthenticated
+
+from profiles_api import permissions
+
+from profiles_api.theory_page.theory_page_serializer import TheoryPageSerializer
+from profiles_api.theory_page.theory_page_model import TheoryPage
+
+
+class TheoryPageViewSet(viewsets.ModelViewSet):
+    """Handles creating, reading and updating theory pages"""
+    authentication_classes = (TokenAuthentication,)
+    serializer_class = TheoryPageSerializer
+    queryset = TheoryPage.objects.all()
+    permission_classes = (permissions.UpdateOwnStatus, IsAuthenticated)
+
+    def perform_create(self, serializer):
+        """Sets the user profile to the logged in user"""
+        serializer.save(user_profile=self.request.user)
+
+
